// test/generated/dist/workingDocumentsCmrForm/workingDocumentsCmrForm.models.ts
export namespace WorkingDocumentsCmrFormModels {
// ...
  export const CmrDocumentResponseDTOSchema = z
    .object({
      id: z.string().describe("ID of the document"),
      name: z.string().describe("Name of the document"),
      positionNumber: z.string().describe("Position number"),
      nameSuffix: z.string().describe("Name suffix of the document").nullish(),
      defaultFileName: z.string(),
      positionId: z.string().describe("Position ID"),
      versionNumber: z.number().describe("Version number of the document"),
      cmrNumber: z.string().describe("CMR number").nullish(),
      consignee: CmrDocumentBusinessPartnerResponseDTOSchema.describe("Consignee information").nullish(),
      shipper: CmrDocumentBusinessPartnerResponseDTOSchema.describe("Shipper information").nullish(),
      carrier: CmrDocumentBusinessPartnerResponseDTOSchema.describe("Carrier information").nullish(),
      successiveCarrier: CmrDocumentBusinessPartnerResponseDTOSchema.describe(
        "Successive carrier information",
      ).nullish(),
      delivery: CmrDocumentCityResponseDTOSchema.describe("Delivery information").nullish(),
      takeover: CmrDocumentCityResponseDTOSchema.describe("Takeover information").nullish(),
      annexedDocuments: z.string().describe("Annexed documents").nullish(),
      carriersObservations: z.string().describe("Carriers observations").nullish(),
      senderInstructions: z.string().describe("Sender instructions").nullish(),
      reimbursement: z.string().describe("Reimbursement information").nullish(),
      specialAgreements: z.string().describe("Special agreements").nullish(),
      suppressWeight: z.boolean().describe("Whether to suppress weight information"),
      suppressMeasurement: z.boolean().describe("Whether to suppress measurement information"),
      freightPaid: z.boolean().describe("Whether freight is paid"),
      freightToBePaid: z.boolean().describe("Whether freight is to be paid"),
      creationPlace: z.string().describe("Place of creation"),
      creationDate: z.string().describe("Date of creation"),
      costsForShipper: CmrDocumentCostsResponseDTOSchema.describe("Costs for shipper").nullish(),
      costsForConsignee: CmrDocumentCostsResponseDTOSchema.describe("Costs for consignee").nullish(),
      selectedCargoIds: z.array(z.string()).readonly().describe("List of selected cargo IDs"),
      cargo: z.array(CmrDocumentCargoResponseDTOSchema).readonly().describe("List of cargo information").nullish(),
      config: CommonModels.HBLDocumentConfigDtoSchema.describe("Configuration settings for the document"),
      bodyRemarks: CommonModels.EditorContentResponseDtoSchema.describe("Body remarks").nullish(),
      footerRemarks: CommonModels.EditorContentResponseDtoSchema.describe("Footer remarks").nullish(),
      settings: CmrDocumentSettingsDtoDTOSchema.describe("Settings for the BL Instructions document").nullish(),
    })
    .readonly();
  export type CmrDocumentResponseDTO = z.infer<typeof CmrDocumentResponseDTOSchema>;
// ...
  export const UpdateCmrDocumentRequestDTOSchema = z
    .object({
      nameSuffix: z.string().describe("Name suffix"),
      cmrNumber: z.string().describe("CMR number"),
      consignee: UpdateCmrDocumentBusinessPartnerRequestDTOSchema.describe("Consignee information"),
      shipper: UpdateCmrDocumentBusinessPartnerRequestDTOSchema.describe("Shipper information"),
      carrier: UpdateCmrDocumentBusinessPartnerRequestDTOSchema.describe("Carrier information"),
      successiveCarrier: UpdateCmrDocumentBusinessPartnerRequestDTOSchema.describe("Successive carrier information"),
      delivery: UpdateCmrDocumentDeliveryRequestDTOSchema.describe("Delivery information"),
      takeover: UpdateCmrDocumentDeliveryRequestDTOSchema.describe("Takeover information"),
      annexedDocuments: z.string().describe("Annexed documents"),
      carriersObservations: z.string().describe("Carriers observations"),
      senderInstructions: z.string().describe("Sender instructions"),
      reimbursement: z.string().describe("Reimbursement"),
      specialAgreements: z.string().describe("Special agreements"),
      suppressWeight: z.boolean().describe("Whether to suppress weight information"),
      suppressMeasurement: z.boolean().describe("Whether to suppress measurement information"),
      freightPaid: z.boolean().describe("Whether freight is paid"),
      freightToBePaid: z.boolean().describe("Whether freight is to be paid"),
      creationPlace: z.string().describe("Place of creation"),
      creationDate: z.string().describe("Date of creation"),
      costsForShipper: UpdateCmrDocumentCostsRequestDTOSchema.describe("Costs for shipper"),
      costsForConsignee: UpdateCmrDocumentCostsRequestDTOSchema.describe("Costs for consignee"),
      selectedCargoIds: z.array(z.string()).readonly().describe("Selected cargo IDs"),
      bodyRemarks: CommonModels.EditorContentUpdateDtoSchema.describe("Body remarks"),
      footerRemarks: CommonModels.EditorContentUpdateDtoSchema.describe("Footer remarks"),
      settings: UpdateCmrDocumentSettingsRequestDTOSchema.describe("Settings"),
    })
    .readonly();
  export type UpdateCmrDocumentRequestDTO = z.infer<typeof UpdateCmrDocumentRequestDTOSchema>;
}